Validate ip and data separately
