+++ packages/p2p/source/socket-server/plugins/validate-data.ts
@@ -3,7 +3,7 @@ import { inject, injectable, tagged } from "@mainsail/container";
 import { Contracts, Identifiers } from "@mainsail/contracts";
 import { Providers } from "@mainsail/kernel";
 
-import { getPeerIp, isValidVersion } from "../../utils/index.js";
+import { isValidVersion } from "../../utils/index.js";
 import {
 	GetApiNodesRoute,
 	GetBlocksRoute,
@@ -19,16 +19,13 @@ import {
 import { BasePlugin } from "./base-plugin.js";
 
 @injectable()
-export class ValidatePlugin extends BasePlugin {
+export class ValidateDataPlugin extends BasePlugin {
 	@inject(Identifiers.Application.Instance)
 	protected readonly app!: Contracts.Kernel.Application;
 
 	@inject(Identifiers.ServiceProvider.Configuration)
 	@tagged("plugin", "p2p")
 	private readonly configuration!: Providers.PluginConfiguration;
-
-	@inject(Identifiers.P2P.Peer.Processor)
-	private readonly peerProcessor!: Contracts.P2P.PeerProcessor;
 
 	public register(server) {
 		if (this.configuration.getRequired("developmentMode.enabled")) {
@@ -50,16 +47,6 @@ export class ValidatePlugin extends BasePlugin {
 
 		server.ext({
 			method: async (request: Contracts.P2P.Request, h: ResponseToolkit) => {
-				const ip = getPeerIp(request);
-
-				if (this.peerDisposer.isBanned(ip)) {
-					return this.banAndReturnBadRequest(request, h, "Validation failed (peer is bannned)");
-				}
-
-				if (!this.peerProcessor.validatePeerIp(ip)) {
-					return this.disposeAndReturnBadRequest(request, h, "Validation failed (bad ip)");
-				}
-
 				const version = request.payload?.headers?.version;
 				if (version && !isValidVersion(this.app, version)) {
 					return this.disposeAndReturnBadRequest(request, h, "Validation failed (invalid version)");
